Now passes a reference and clones on-demand instead of cloning at each call (like an idiot), leading to a ~5x speedup


Time to calculate N!:

N | Previous| New
-- | -- | --
0 | 5ms | 1ms
1 | 7ms | 3ms
2 | 94ms | 13ms
3 | 819ms | 181ms
4 | 10.750s | 2.417s
5 | 133.489s | 25.427s